requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
edgartools==5.8.3
filelock==3.20.2
google-auth==2.47.0
google-genai==1.57.0
h11==0.16.0
hishel==0.1.3
httpcore==1.0.9
httpx==0.28.1
httpxthrottlecache==0.3.0
humanize==4.15.0
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.4.0
orjson==3.11.5
pandas==2.3.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyrate-limiter==3.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
rank-bm25==0.2.2
RapidFuzz==3.14.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
stamina==25.2.0
tabulate==0.9.0
tenacity==9.1.2
textdistance==4.6.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.3
websockets==15.0.1
